requirements.txt
social-auth-app-django==1.2.0
pillow==4.3.0
mysqlclient==1.3.9
ipython==5.5.0
django-webpack-loader==0.5.0
django-rest-framework==0.1.0
django-debug-toolbar==1.6
